inverse int et float

+++ python_variable_annotations/7-to_kv.py
@@ -3,6 +3,6 @@
 from typing import Tuple, Union
 
 
-def to_kv(k: str, v: Union[float, int]) -> Tuple[str, float]:
+def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
     """ return str k and the square float of v """
-    return (k, v ** 2)
+    return (k, float(v ** 2))
